form validation, default select value and confirmation dialog added

+++ client/src/components/MainComponent.vue
@@ -16,7 +16,7 @@
             <td>{{rout.day_of_week}}</td>
             <td>{{rout.visit_date}}</td>
             <td><button v-on:click="getRoutUrl(rout.trade_point_id)">Show on Map</button></td>
-            <td><button class="btn btn-warning" v-on:click="deleteRouteClicked(rout.id)">Delete</button></td>
+            <td><button class="btn btn-danger" v-on:click="deleteRouteClicked(rout.id)">Delete</button></td>
             <td><button class="btn btn-success" v-on:click="editRout(rout)">Update</button></td>
           </tr>
         </tbody>
@@ -36,26 +36,36 @@
           <input type="hidden" :value="rout.id" >
           <div class="form-group">
             <select v-model="selected_point">
-              <option disabled value="">Выберите точку поставки</option>
+              <option :value="null" disabled>Выберите точку поставки</option>
               <option v-for="point in points" v-bind:value="point.id" v-bind:key="point.id">{{point.name}}</option>
             </select>
           </div>
 
           <div class="form-group">
               <select v-model="selected_agent">
-                <option disabled value="">Выберите агента</option>
+                <option :value="null" disabled>Выберите агента</option>
                 <option v-for="agent in agents" v-bind:value="agent.id" v-bind:key="agent.id">{{agent.first_name}}</option>
               </select>
           </div>
+          <template>
+            <div style="display: flex; flex-direction: row; justify-content: space-between;">
+              <input type="radio" id="one" value="Назначить на день недели" v-model="picked">
+            <label for="one" style="margin-top: -6px;">Назначить на день недели</label>
+            <br>
+            <input type="radio" id="two" value="Назанчить на конкретную дату" v-model="picked">
+            <label for="two" style="margin-top: -6px;">Назанчить на конкретную дату</label>
+            <br>
+            </div>
+          </template>
 
-          <div class="form-group">
+          <div class="form-group" v-if="picked === 'Назначить на день недели'">
               <select v-model="selected_day">
-                <option disabled value="">Выберите день недели</option>
+                <option :value="null" disabled>Выберите день недели</option>
                 <option v-for="day in days" v-bind:value="day.value" v-bind:key="day.label">{{day.label}}</option>
               </select>
           </div>
 
-          <div class="form-group">
+          <div class="form-group" v-if="picked === 'Назанчить на конкретную дату'">
             <span>Выберите день поставки</span>
             <input v-model="visit_date"  type="date" name="date" 
             data-date-format="YYYY-MM-DD" value="2019-12-15"
@@ -76,6 +86,7 @@
 <script>
 import MainDataService from '../services/MainDataService';
 import { Form } from 'vform';
+
 
 window.Form = Form;
 
@@ -109,6 +120,7 @@ export default {
       agents: [],
       selected_agent: null,
       visit_date: null,
+      picked: 'Назначить на день недели',
       form: new Form({
                     id:'',
                     trade_point_id : '',
@@ -135,7 +147,10 @@ export default {
   },
     newModal(){
       this.editmode = false;
-      this.form.reset();
+      this.selected_point = null;
+      this.selected_agent = null;
+      this.selected_day = null;
+      this.visit_date = null;
       this.$modal.show('points');
   },
   editRout(rout) {
@@ -171,17 +186,41 @@ export default {
                     newRout.agent_id = this.selected_agent;
                     newRout.day_of_week = this.selected_day;
                     newRout.visit_date = this.visit_date;
-                MainDataService.createRout(newRout)
-                .then(()=>{
-                  this.$modal.hide('points');
-                  this.refreshRoutes();
-                })
+                    if(newRout.visit_date) newRout.trade_point_id = null;
+                    if(newRout.trade_point_id && newRout.agent_id && (newRout.day_of_week || newRout.visit_date)) {
+                      MainDataService.createRout(newRout).then(() => {
+                        this.$modal.hide('points');
+                        this.refreshRoutes();
+                      })
+                    } else {
+                      alert('Заполены не все поля');
+                    }
+                
     },
   deleteRouteClicked(id) {
-      MainDataService.deleteRouts(id).then(() => {
-        this.message = `Delete of course ${id} Successful`;
-        this.refreshRoutes();
-      });
+     this.$swal({
+                    title: 'Are you sure?',
+                    text: "You won't be able to revert this!",
+                    type: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#3085d6',
+                    cancelButtonColor: '#d33',
+                    confirmButtonText: 'Yes, delete it!'
+                    }).then((result) => {
+                        // Send request to the server
+                         if (result.value) {
+                                 MainDataService.deleteRouts(id).then(() =>{
+                                        this.$swal(
+                                        'Deleted!',
+                                        'Your file has been deleted.',
+                                        'success'
+                                        )
+                                    this.refreshRoutes();
+                                }).catch(()=> {
+                                    this.$swal("Failed!", "There was something wronge.", "warning");
+                                });
+                         }
+                    })
     },
     getRoutUrl(pointId) {
       MainDataService.retrieveAllPoints()
@@ -202,4 +241,9 @@ export default {
 </script>
 
 <style>
+.v--modal-box.v--modal {
+  width: 500px!important;
+  height: 310px!important;
+  padding: 10px;
+}
 </style>
